added four PLOS journals and PLOS organization
Note that I formatted the individual PLOS journals as separate entities in the journal list (unlike, for example, the journals nested under AAAS) because not all of the PLOS journals have signed on together. The umbrella organization has signed on, so I also added PLOS in the organization signatories. This is the format that PLOS agreed to and will be updated if/when the other journals sign on.