small edits to tweet text
I made one of the 3 tweets include a link to the youtube video https://youtu.be/SWkqdNppL-s and made it so that all 3 had the OpenScience hashtag. I removed the OpenData hashtag because we've already had a bit of confusion from some entrants wondering if we required open data as part of this competition, we do not.